requirements-rag.txt
chromadb
openai
python-dotenv
numpy
